requirements.txt
julius==0.2.7
ml_collections==0.1.1
numpy
PyYAML
soundfile==0.12.1
torch>=2.2.0
torchaudio>=2.2.0
tqdm
einops==0.6.1
